Move files to volume folder.
