+++ util/test/test_cleancase.py
@@ -1,7 +1,7 @@
-from enforce_typing import enforce_types
 import pytest
+from enforce_typing import enforce_types
 
-from util import cleancase
+from util.volume import cleancase
 
 
 @enforce_types
