+++ util/volume/allocations.py
@@ -1,6 +1,7 @@
 from enforce_typing import enforce_types
 
-from util import cleancase, csvs
+from util import csvs
+from util.volume import cleancase
 
 
 @enforce_types
