+++ util/volume/calcrewards.py
@@ -1,12 +1,13 @@
 from datetime import datetime
 from typing import Dict, List, Tuple, Union
 
-from enforce_typing import enforce_types
 import numpy as np
 import scipy
+from enforce_typing import enforce_types
 
-from util import cleancase as cc, tousd
 from util.constants import MAX_N_RANK_ASSETS, RANK_SCALE_OP
+from util.volume import cleancase as cc
+from util.volume import tousd
 
 # Weekly Percent Yield needs to be 1.5717%., for max APY of 125%
 TARGET_WPY = 0.015717
